Use int64_t for chunk_size in foreach kernels (#3759)
Fix potential integer overflow in foreach kernels for large tensors.
Widen `chunk_size` parameters from `int` to `int64_t` in XPU
foreach/fused optimizer functors to prevent overflow when processing
tensors >2^31 elements.

It's a cherry-pick of https://github.com/pytorch/pytorch/pull/156872

+++ src/ATen/native/xpu/sycl/ForeachFunctors.h
@@ -226,7 +226,7 @@ struct PointwiseOpScalarFunctor {
   using opmath_t = at::opmath_type<T>;
   template <typename Op, typename TLA, typename TLW>
   void operator()(
-      const int chunk_size,
+      const int64_t chunk_size,
       TLA tlAddress,
       TLW tlWGMeta,
       sycl::nd_item<1> item_id,
@@ -292,7 +292,7 @@ struct PointwiseOpScalarListFunctor {
   using opmath_t = at::opmath_type<T>;
   template <typename Op, typename TLA, typename TLW>
   void operator()(
-      const int chunk_size,
+      const int64_t chunk_size,
       TLA tlAddress,
       TLW tlWGMeta,
       sycl::nd_item<1> item_id,
@@ -415,7 +415,7 @@ void binary_op_scalar(
     T** args,
     opmath_t scalar,
     int n,
-    int chunk_size,
+    int64_t chunk_size,
     bool all_aligned,
     Op op,
     size_t item_range,
@@ -472,7 +472,7 @@ void binary_op_scalar_tensor(
     scalar_t* scalar,
     opmath_t alpha,
     int n,
-    int chunk_size,
+    int64_t chunk_size,
     bool all_aligned,
     Op op,
     size_t item_range,
@@ -631,7 +631,7 @@ struct BinaryOpScalarListFunctor {
   using opmath_t = at::opmath_type<T>;
   template <typename TLA, typename TLW, typename Op>
   void operator()(
-      int chunk_size,
+      int64_t chunk_size,
       TLA tlAddress,
       TLW tlWGMeta,
       sycl::nd_item<1> item_id,
@@ -706,7 +706,7 @@ struct TernaryOpListFunctor {
   using opmath_t = at::opmath_type<T>;
   template <typename TLA, typename TLW, typename Op>
   void operator()(
-      int chunk_size,
+      int64_t chunk_size,
       TLA tlAddress,
       TLW tlWGMeta,
       sycl::nd_item<1> item_id,
@@ -773,7 +773,7 @@ struct TernaryOpScalarFunctor {
   using opmath_t = at::opmath_type<T>;
   template <typename TLA, typename TLW, typename Op>
   void operator()(
-      int chunk_size,
+      int64_t chunk_size,
       TLA tlAddress,
       TLW tlWGMeta,
       sycl::nd_item<1> item_id,
@@ -844,7 +844,7 @@ struct TernaryOpScalarListFunctor {
 
   template <typename TLA, typename TLW, typename Op>
   void operator()(
-      int chunk_size,
+      int64_t chunk_size,
       TLA tlAddress,
       TLW tlWGMeta,
       sycl::nd_item<1> item_id,
